Implemented 5 minute timer to turn off the service.
Timer is restarted when the service is used, so it stops 5 minutes
after the final use.

Possibly if someone has an open connection to the service, it won't
stop.  Probably should find a way to correct that to force a client of
the service to re-authenticate after doing nothing beyond the
timeout. Not positive, need to test.