Experimental code for "GET_PASSWORD". It fetches a password based on a content URI which must correspond to a category and a description. The category must match the package of the caller (not sure if this is an effective enforcement mechanism), so that the caller can only get & set their own passwords:
content://org.openintents.keys/org.syntaxpolice.ServiceTest/opensocial
^ scheme  ^host                ^ category=caller            ^ description

The ServiceTest application has also been extended to try this out. As
an example, you can add a category of "org.syntaxpolice.ServiceTest"
(which will correspond to the test application) and several passwords,
one of which should have the description "opensocial".  Use the 'get'
button in the test application to pull out the username & password.

Obviously more to be done here, but it's a start, and I'd be
interested in feedback.