Added SetPassword implementation.
Open question: Should we provide extras for all of the fields for a
password entry?

Modified addCategory to dbHelper to NOT add a category if one with the
same name already exists. This simplified the logic in some spots.

Test driver has new fields to set description / usernmae /
password. You can use these with 'set' button to add a password, then
query it with "get".